+++ decidim-core/lib/decidim/core/test.rb
@@ -77,4 +77,4 @@ require "decidim/core/test/shared_examples/digest_mail_examples"
 require "decidim/core/test/shared_examples/hideable_resource_examples"
 require "decidim/core/test/shared_examples/active_support_examples"
 require "decidim/core/test/shared_examples/statistics_cell_examples"
-require "decidim/core/test/shared_examples/resource_locator_presener_examples"
+require "decidim/core/test/shared_examples/resource_locator_presenter_examples"
