+++ decidim-core/lib/decidim/core/test/shared_examples/resource_locator_presenter_examples.rb
